Moves View to testsite
- Updated query
- Updated template name + references
