+++ testsite/templates/index.html
@@ -24,6 +24,6 @@
         <li><a href="/editables/">Editables</a></li>
       </ul>
   </div>
-  {% include "pages/app/news_feed.html" %}
+  {% include "pages/app/_news_feed.html" %}
 </div>
 {% endblock %}
